multithreaded based and general refactor
