+++ requirements.txt
@@ -1,5 +1,5 @@
 aiohttp
 aiohttp_retry
-runapy==0.7.2
+runapy
 requests>=2.32.0
 pydantic>=2.8.2
