+++ test_data/project_allocations_02-18-25_to_02-25-25.csv
@@ -1,0 +1,30 @@
+Department,Project,Project Allocated GPUs,Allocated GPU - Peak,Allocated GPU - Avg,CPU Memory (GB) - Peak,CPU Memory (GB) - Avg,CPU (# Cores) - Peak,CPU (# Cores) - Avg
+Unknown,test,0.00,0.00,0.00,0.23,0.03,1.00,0.11
+Unknown,team-a,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,test-notificiation-works,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,proj1,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,blatest,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,zurb,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,sheeesh,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,clearml,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,project-f,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,qer,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,p-1,0.00,0.00,0.00,0.15,0.08,1.00,0.46
+Unknown,repro-proj,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,testproj01,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,bbb,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,zzz,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,pop,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,argo-wf,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,zombie,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,test-fractions,1.00,1.00,0.50,0.22,0.20,0.81,0.42
+Unknown,proj-pol,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,single,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,p3,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,foofoofafa,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,test1,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,project-d,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,itayq,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,proj2,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,w4t,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+Unknown,test2,0.00,0.00,0.00,0.00,0.00,0.00,0.00
