+++ test_data/utilization_metrics_02-18-25_to_02-25-25.csv
@@ -1,0 +1,51 @@
+Project,User,Job Name,GPU Hours,Memory Hours (GB),CPU Hours,GPU Allocated,GPU Utilization % - Peak,GPU Utilization % - Average,CPU Utilization (Cores) - Peak,CPU Utilization (Cores) - Average,CPU Memory (GB) - Peak,CPU Memory (GB) - Average
+test,devops@run.ai,mlf-v2,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+team-a,devops@run.ai,tensor4,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+test-notificiation-works,snir.shrim@run.ai,test1,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+proj1,devops@run.ai,example1,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+blatest,devops@run.ai,job-70f933f1d00d,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+zurb,Unknown,throw-u-off-v3,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+sheeesh,devops@run.ai,vscode-rewrite,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+clearml,Unknown,clearml-id-f3fdbe1b0c71421cbbbc0ffb19d90752,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+project-f,cs-demo-api,gpu-2-workload-h,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+qer,devops@run.ai,blurb,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+test,devops@run.ai,fail-oom-v2,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+team-a,devops@run.ai,fdsfdsa,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+p-1,devops@run.ai,dist-1,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+repro-proj,app_admin_v1@itay.xyz,repro-train,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+testproj01,devops@run.ai,training-860db51d2bc4,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+bbb,Unknown,check999,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+test,devops@run.ai,gpu-only,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+project-f,cs-demo-api,cpu-6-workload-i,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+zzz,devops@run.ai,testme-no-tools-v3-with-data,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+zzz,devops@run.ai,testme-no-tools,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+test,devops@run.ai,mlf-v5,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+pop,Unknown,throw-u-off-v3,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+argo-wf,Unknown,itay-workflow-wawawa,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+zombie,devops@run.ai,env-lower,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+test-fractions,devops@run.ai,test-multi-secret-2,0.00,0.00,0.07,0.00,0.00,0.00,0.81,0.80,0.19,0.19
+proj-pol,devops@run.ai,dsfh,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+single,devops@run.ai,test-2cpu-v2,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+p3,devops@run.ai,t8,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+test-fractions,devops@run.ai,test-jupyter,0.09,0.00,0.00,1.00,0.00,0.00,0.03,0.03,0.22,0.22
+foofoofafa,devops@run.ai,ubuntu-vs,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+test,devops@run.ai,test-jupyter,0.00,0.00,14.12,0.00,0.00,0.00,1.00,0.95,0.23,0.23
+test,devops@run.ai,hello123-b2cbf935c5d6,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+single,devops@run.ai,test-1cpu,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+argo-wf,Unknown,itay-workflow-five,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+single,devops@run.ai,test-1cpu-should-fail-v2,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+test1,ofir.eldar,job5,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+project-d,cs-demo-api,cpu-4-workload-f,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+itayq,devops@run.ai,test-my-nfs,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+test,devops@run.ai,pi-cliv2-test-2,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+foofoofafa,devops@run.ai,vsco2,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+testproj01,devops@run.ai,training-f1dc51edd5f8,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+test,Unknown,rayjobomritest,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+argo-wf,devops@run.ai,this-must-shutdown,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+testproj01,devops@run.ai,training-57caab638654,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+proj2,devops@run.ai,sdfd,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+proj2,devops@run.ai,dsf,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+test,devops@run.ai,test-gpu-before,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+w4t,devops@run.ai,shugtest,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+test2,devops@run.ai,dsdsa,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00,0.00
+p-1,devops@run.ai,ws-1,0.00,0.00,150.15,0.00,0.00,0.00,1.00,0.91,0.15,0.15
